+++ NEWS.md
@@ -14,6 +14,9 @@ multiFeature plots as well.
 * Fix bug in `getResultsUpset()` that failed when using the absolute value
 filters (`|<|`, `|>|`)
 
+* `getResultsUpset()` now throws an error when there are no features remaining
+for plotting after the filters are applied
+
 # 0.23.1
 
 * The release tarball includes version 0.3.7 of the web app